backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
openai
python-dotenv
